conn: move raw connection to its own file
